Add USDR repo basics and Python dependency list (#2)
This is just adding a few basic bits we want most public repos at USDR to have:

- License info
- Code of Conduct
- Contributing Guide

Plus a `requirements.txt` file for the Python dependencies, so it’s easier for someone to see what they need to install (and do so if using pip). This should also let GH give us automated PRs for security updates.