Add script for counting total seen slots (#3)
We’ve had a few requests over time for a total count of slots that UNIVAF has monitored, so I thought I’d clean up the hacky script I originally wrote for this and add it here for later re-use. This also adds support for gzipped NDJSON files, since that is the format we are now writing for all log files.